🚚 rename: userInfoPopup 분리 및 닉네임 최대글자 설정 #62
